functions/api/requirements.txt
google-cloud-firestore
flask==2.3.3
functions_framework
